// lib/api/types.gen.ts
export type AccountCashEquivalent = {
  assetType: 'CASH_EQUIVALENT'
  cusip?: string | null
  symbol?: string | null
  description?: string | null
  instrumentId?: number | null
  netChange?: number | null
  type?: Type11 | null
}

export type AccountCollectiveInvestment = {
  assetType: 'COLLECTIVE_INVESTMENT'
  cusip?: string | null
  symbol?: string | null
  description?: string | null
  instrumentId?: number | null
  netChange?: number | null
  type?: Type2 | null
}

export type AccountEquity = {
  assetType: 'EQUITY'
  cusip?: string | null
  symbol?: string | null
  description?: string | null
  instrumentId?: number | null
  netChange?: number | null
}

export type AccountFixedIncome = {
  assetType: 'FIXED_INCOME'
  cusip?: string | null
  symbol?: string | null
  description?: string | null
  instrumentId?: number | null
  netChange?: number | null
  maturityDate?: string | null
  factor?: number | null
  variableRate?: number | null
}

export type AccountMutualFund = {
  assetType: 'MUTUAL_FUND'
  cusip?: string | null
  symbol?: string | null
  description?: string | null
  instrumentId?: number | null
  netChange?: number | null
}

export type AccountOption = {
  assetType: 'OPTION'
  cusip?: string | null
  symbol?: string | null
  description?: string | null
  instrumentId?: number | null
  netChange?: number | null
  optionDeliverables?: Array<AccountAPIOptionDeliverable> | null
  putCall?: PutCall | null
  optionMultiplier?: number | null
  type?: Type12 | null
  underlyingSymbol?: string | null
}

export type AccountsInstrument =
  | AccountCashEquivalent
  | AccountEquity
  | AccountFixedIncome
  | AccountMutualFund
  | AccountOption
  | AccountCollectiveInvestment
// ...
export type Position = {
  shortQuantity?: number | null
  averagePrice?: number | null
  currentDayProfitLoss?: number | null
  currentDayProfitLossPercentage?: number | null
  longQuantity?: number | null
  settledLongQuantity?: number | null
  settledShortQuantity?: number | null
  agedQuantity?: number | null
  instrument?: AccountsInstrument | null
  marketValue?: number | null
  maintenanceRequirement?: number | null
  averageLongPrice?: number | null
  averageShortPrice?: number | null
  taxLotAverageLongPrice?: number | null
  taxLotAverageShortPrice?: number | null
  longOpenProfitLoss?: number | null
  shortOpenProfitLoss?: number | null
  previousSessionLongQuantity?: number | null
  previousSessionShortQuantity?: number | null
  currentDayCost?: number | null
}